Added eisen version and author Added host and group variable list API Added endpoint for group variable list and host variable list Added set Group variable Added port to host_fields
We still didn't added Eisen version (0.0.1) and author ('Eisen'), this variable are needed for be used in the setup.py
We also needed to add and change target group and target host variables, for example password or ssh port. so we added the API variable list and endpoint at (hostvars and groupvars).
  for doing this we added a set Group Variable function with the already present set host variable function. Because function can be set at host and at group.
We also needed to add the target host port directly bypassing the variable setter because is one of the "essential" setting.