Fix #9 Adding variable with post in /host/<id>/vars
Using a global dynamic inventory we can add the variable to the dynamic inventory.
Added test case for check dynamic inventory.

Returning error message if it cannot retrive variable instead of crashing.

Added self to the celery task useful for check the id of the queue instance for multiqueue on same task.