added sqlalchemy root resource package_retrieve resource os_check resource
Because we are moving the long process database manipulation to
python connecting to the eisen frontend database in remote, making
manipulation usually many time more fast than using php5 or php7 and
also not blocking process.
we can also decide to implement multiprocess or similar.
for database manipulation we mean the package list update and operating system check.
this work as been moved from eisen frontend to the eisen agent.