Added commit result directly to db using sqlalchemy and implemented packages resource for installing package.
Because php is consuming too much time updating package to the database, we moved
all the DB write part directly into the Eisen Agent, this is making more fast the commiting fase for many packages.
We are still trying to merge celery and sqlalchemy for making it completely not blocking.
but this is anyway a starting and working point.