Packaged eisen_engine for python setuptools
Because installing eisen_engine is difficult as now,
we are moving for adding it to pip and making it more
simple to install and use.
This patch is making the setup.py work with setuptools,
so as now eisen_engine can be installed by using
python setup.py install.
the configuration file is under ~/.eisen_engine.conf
the runner file is eisen_engine